requirements.txt
Django
django-bootstrap3
django-recaptcha